refactor: extract factory module
